Added theme and locales buttons to the login page. Made a component for the theme button.
